Move classes for typing to types.py
